Changed how local is pulled out of the fullpath
Assumes that the first directory after resources/lang is the local name.  This means if you maintain a full directory structure under lang/[local] you will not have a problem sorting out localization files.
Prior to this the local was found by getting the parent folder name of the localization files path.  This only works if all of your localization scripts are right under the local (en/fr) folder.